fix don't delete item on backspace in input field
When deleting the current quanity by hitting backspace in the input field, it would trigger a change event with the target value being 0.
For quantity of zero an item would get deleted. So when a user wants to enter in an amount by first deleting the current quantity then the whole item would be removed from the cart.